Move the metrics code into a module
